feat: complete VMO syscalls, kill old DMA/memory_share, renumber (v0.6 Phase 3a, Cluster B)
Phase 3a Cluster B: snapshot/restore (syscalls 21-22), seal (23),
op_range with COMMIT/DECOMMIT/LOOKUP (24), mapping tracking for
cross-process PTE invalidation, sealed-aware fault handler,
PermissionDenied error, snapshot_count in VmoInfo.

Kill the old way: dma_alloc, dma_free, memory_share kernel syscalls
removed. sys library provides identical API signatures backed by VMOs
internally (contiguous VMO + cross-process vmo_map + vmo_op_range
LOOKUP for PA retrieval). No userspace service code changed.

Syscall renumber: dense 0-36 grouped by abstraction layer (runtime,
capability, IPC, sync, time, heap, VMO, process/thread, scheduling,
device). No gaps, no implementation-order artifacts. Syscall filter
mask widened from u32 to u64 to cover full table.

37 syscalls, 2462 tests pass.

+++ system/kernel/process.rs
@@ -37,9 +37,9 @@ pub struct Process {
     /// last running thread is rescheduled away.
     pub(crate) killed: bool,
     /// Per-process syscall filter bitmask. Bit N set = syscall N allowed.
-    /// Defaults to u32::MAX (all syscalls allowed). Set before process_start
+    /// Defaults to u64::MAX (all syscalls allowed). Set before process_start
     /// via the PROCESS_SET_SYSCALL_FILTER syscall.
-    pub(crate) syscall_mask: u32,
+    pub(crate) syscall_mask: u64,
 }
 /// Unique process identifier. Index into the scheduler's process table.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -53,7 +53,7 @@ impl Process {
             thread_count: 0,
             started: false,
             killed: false,
-            syscall_mask: u32::MAX,
+            syscall_mask: u64::MAX,
         }
     }
 }
